semantic_kernel_agent/requirements.txt
semantic-kernel>=1.36.0
python-dotenv